Update README.md
This is a web based math game for kids. It takes the number drawn in a black box as an input and makes a prediction of which number that is. The Tensorflow model that processes the input image was created in Python.

You can access to the web page and try this game with the following link: https://begies-projects.github.io/funny-math/